Qair-package(Qair)	R Documentation

Traitement de donnees domaine de la qualite de l'air

Description

Ce package definit principalement deux types d'objet qui permettent, via les fonctions definies, de traiter les donnees de qualite de l'air de maniere temporelle et par rapport a des donnees de vent. Les fonctions permettent de transformer des donnees d'une base temporelle a une autre, de calculer des profils (journaliers, hebdomadaires, mensuels, annuels) et des roses de pollution.

Details

Package: 	 Qair

Type: 	 Package

Version: 	 0.16.4

Date: 	 2008-10-13

License: 	 GPL

Author(s)

Responsable : Vladislav Navel <vladislav.navel@atmo-poitou-charentes.org>

[Package Qair version 0.16.4 Index]