Xair2R(Qair)	R Documentation

Extraire des donnees d'une base Xair.

Description

Dans le domaine de la qualite de l'air, les donnees sont souvent stockees dans une base Xair. Cette fonction permet d'y acceder simplement.

Usage

Xair2R(polluants, dated, datef, dt=c("qh", "heure", "jour", "mois", "an"), merging, codeV = c("A", "R", "O", "W"), dsn=NULL, uid=NULL, pwd=NULL, brute=FALSE)

Arguments

polluants 	 Noms des polluants a aller chercher dans la base.

dated 	 date de debut au format "y-m-d"

datef 	 date de fin au format "y-m-d"

dt 	 Base temporelle des donnees a extraire.

merging 	 Un 'Qair' deja existant auquel ajouter les nouvelles donnees

codeV 	 Vecteur de 'character' indiquant les codes de validation de Xair pour lesquels les donnees doivent etre rapatriees.

dsn 	 Nom de la base de donnees Xair telle que definie dans le pilote ODBC. Si NULL, la fonction demandera a chaque premiere utilisation une valeur, puis la conservera dans options('Xair.dsn').

uid 	 Identifiant a utiliser la connection a Xair. Si NULL, la fonction demandera a chaque premiere utilisation une valeur, puis la conservera dans options('Xair.uid').

pwd 	 Mot de passe pour la connection a Xair. Si NULL, la fonction demandera a chaque premiere utilisation une valeur, puis la conservera dans options('Xair.pwd').

brute 	 'logical' indiquant si les donnees a charger son les donnees brutes ou les donnees validees. (n'est utilise que pour les donnees qh, horaires et journalieres.

Value

Renvoie un objet de type 'Qair'.

Examples

#  # ne fonctionne que si une 'data source name' appelee BaseN correspond a une base Xair, et 
#  # si des mesures dont le nom court de mesure est l'un de c("O3_VAU", "O3_VER", "O3_AYT"), sont definies dans la base. 
#  donnees <- Xair2R(c("O3_VER", "N2_VER", "NO_VER", "O3_VAU", "O3_AYT", "DV_ATM", "VV_ATM"), "07-01-01", "07-12-31", "qh")  
#  head(donnees) 
#  mesures(donnees) 
#  unites(donnees) 
#  plot(donnees) 

[Package Qair version 0.16.4 Index]