as.data.frame.Qair(Qair)	R Documentation

Transformation d'un objet Qair en objet data.frame

Description

Permet d'obtenir une data.frame a partir d'un objet de la classe Qair

Usage

  ## S3 method for class 'Qair':
  as.data.frame(x, row.names = NULL, optional = FALSE, ...)

Arguments

x 	 une structure de donnees dont la classe herite de "Qair". Objet renvoye par l'une des fonctions suivantes Xair2R, toHeure, toJour, toHebdo, toMois, toAn, toPJour, toPHebdo, toPMois, toPAn

row.names 	 inutilise. Necessaire pour la compatibilte avec la methode generic

optional 	 inutilise. Necessaire pour la compatibilte avec la methode generic

... 	 inutilise. Necessaire pour la compatibilte avec la methode generic

Value

Cette fonction est essentiellement utile en interne pour l'application des fonctions recurrentes sur la classe Qair. Elle renvoie le tableau de donnees contenu dans "donnees" en supprimant tous les attributs supplementaires caraterisant la classe Qair.

See Also

Xair2R, toHeure, toJour, toHebdo, toMois, toAn, toPJour, toPHebdo, toPMois, toPAn

Examples

  data(larochelle)
  summary(larochelle)
  class(larochelle)
  class(as.data.frame(larochelle))

[Package Qair version 0.16.0 Index]