calcRose(Qair)	R Documentation

Calcul sur les donnees pour obtenir des roses de pollution

Description

Permet de realiser les calculs necessaires a la realisation de roses de pollution.

Usage

calcRose(donnees, ...)
## S3 method for class 'Qair':
calcRose(donnees, formula, breaksVent = 0:36 * 10 - 5, FUN, ...)
## S3 method for class 'rose':
calcRose(donnees, FUN, ...)

Arguments

donnees 	 Un element de type 'Qair' contenant une colonne avec des valeurs de vitesses de vent et une colonne avec des valeurs de directions de vent. Ou un element de type 'rose' non encore aggrege. (cf. details).

formula 	 de la forme ~dv+vv avec dv le nom de la colonne des directions de vent et vv celui de la colonne des vitesses de vent.

breaksVent 	 un vecteur d'angles (en degre) correspondant aux separations entre les secteurs souhaites sur la rose de pollution. Les valeurs entrees sont automatiquement ramenees entre 0 et 360 degre.

FUN 	 la fonction a utiliser pour l'aggregation des donnees. Si 'FUN' n'est pas precisee, la fonction 'mean' est appliquee par defaut avec l'argument na.rm=T.

... 	 Permet de preciser des arguments specifiques a la fonction FUN.

Details

La fonction 'calcRose' appliquee a un objet de type 'rose' (cette utilisation est uniquement accessible par les programmes) ou a un objet de type 'Qair' renvoie un objet de type 'rose' aggrege et directement representable via la fonction plot.rose.

Value

Dans le cas d'un objet 'Qair' renvoie un objet 'rose', aggrege. Dans le cas d'un objet 'rose' renvoie un objet 'rose', aggrege. (inaccessible)

See Also

plot.rose, as.list.rose

Examples

  data(larochelle)

  # 1
  calcRose(larochelle, ~DV_ATM+VV_ATM)
  # 2
  calcRose(larochelle, ~DV_ATM+VV_ATM, FUN="mean", na.rm=TRUE)
  # 3
  as.list(calcRose(larochelle, ~DV_ATM+VV_ATM))

  calcRose(larochelle, ~DV_ATM+VV_ATM, breaksVent=c(0, 90, 180, 270), FUN="toPJour", na.rm=TRUE)
  # 4
  as.list(calcRose(larochelle, ~DV_ATM+VV_ATM, breaksVent=c(0, 90, 180, 270), FUN="toPJour", na.rm=TRUE))

[Package Qair version 0.16.0 Index]