couleurs(Qair)	R Documentation

Divers fonctions liees aux couleurs

Description

Chaque fonction retourne des details dans un type de ton.

Usage

chaud(n = 100, inverted = FALSE)
froid(n = 100, inverted = FALSE)
arcenciel(n = 101, inverted = FALSE)

Arguments

n 	 nombre de couleurs souhaitees. (entre 1 et 100, 101 pour la fonction arcenciel)

inverted 	 logical. cf value.

Value

Renvoie un vecteur de couleurs. Si inverted == F, les plus pales sont renvoyees au debut, si inverted == T, les plus sombres sont renvoyees au debut.

Examples

  par(mfrow=c(1, 2))
  plot(NA, xlim=c(0, 100), ylim=c(0, 100))
  abline(v=0:100, col=c(froid(50, TRUE), "white", chaud(50)), lwd=5)
  plot(NA, xlim=c(0, 100), ylim=c(0, 100))
  abline(h=0:100, col=arcenciel(101), lwd=5)

[Package Qair version 0.16.0 Index]