echelle(Qair)	R Documentation

Affecter une couleur selon une echelle de valeur.

Description

Permet d'attribuer a des valeurs numeriques les couleurs associees dans une echelle de couleurs.

Usage

echelle(donnees, mini = 0, maxi = ceiling(max(donnees, na.rm = TRUE)), nclass = 101, inverted = FALSE, col = arcenciel(nclass, inverted))

Arguments

donnees 	 Vecteur de numeric contenant les valeurs auxquelles on souhaite asoocier les couleurs.

mini 	 Valeur minimale de l'echelle de couleur. Par defaut = 0.

maxi 	 Valeur maximale de l'echelle de couleur. Par defaut = valeur maxi de 'donnees'.

nclass 	 nombre de couleurs differentes dans l'echelle de couleurs

inverted 	 Les couleurs de l'echelle doivent-elles etre inversees ?

col 	 Vecteur de couleurs parmi lesquelles les couleurs de l'echelle vont etre selectionnees.

Value

Retourne un vecteur de couleurs de meme taille que donnees.

See Also

froid, chaud, colors

Examples

  plot(0:101, y <- c(0:50, 50:0), type="h", col=echelle(y, 0, 50, 51), lwd=5)

[Package Qair version 0.16.0 Index]