plot.Qair(Qair)	R Documentation

Affiche les donnees d'objet de la classe 'Qair'

Description

La fonction plot definit un format par defaut pour les objets de la classe 'Qair', qu'ils soient des series temporelles ou des profils.

Usage

## S3 method for class 'Qair':
plot(x, y, xlim, ylim, ..., axes=TRUE, x.leg, y.leg, cex.leg=0.7, legend=c("abrev", "complet", "none"), ask.leg=T, out.format="y-m-d")
## S3 method for class 'Qair':
points(x, which = names(x), col, ...)
## S3 method for class 'Qair':
lines(x, which = names(x), col, ...)

Arguments

x 	 Objet de class 'Qair' contenant les donnees a tracer.

y 	 vecteur de 'character' correspondant aux colonnes a afficher.

xlim 	 Limites en abscisse du graphique.

ylim 	 Limites en ordonnees du graphique.

axes 	 logical, les axes doivent-ils etre traces ?

x.leg 	 bord gauche de la legende.

y.leg 	 bord superieur de la legende

cex.leg 	 taille de la police de la legende

legend 	 La legende doit-elle prendre les noms courts des colonnes ('abrev'), les designations completes des mesures ('complet'), ou ne pas etre affichee ('none').

ask.leg 	 logical, l'emplacement de la legende doit-il etre indique par un clique sur le graphe ?

out.format 	 chaine de characteres indiquant le format souhaite pour l'affichage des dates sur le graphique. (cf. chron)

which 	 vecteur de 'character' correspondant aux colonnes a afficher.

col 	 couleurs a affecter aux valeurs devant etre representees.

... 	 Tout argument pouvant etre transmis aux fonctions plot.default, lines.default ou points.default.

See Also

plot.default, lines.default ou points.default

Examples

  data(larochelle)

  # 1{
  par(mfcol=c(3, 2))
  plot(larochelle, grep("O3", names(larochelle), value=TRUE), ask.leg=FALSE)
  plot(toPJour(larochelle), grep("O3", names(larochelle), value=TRUE), legend="complet", ask.leg=FALSE)
  plot(toPJour(larochelle, FUN="max", na.rm=TRUE), grep("O3", names(larochelle), value=TRUE), legend="complet", ask.leg=FALSE)
  plot(larochelle, grep("N", names(larochelle), value=TRUE), ask.leg=FALSE)
  plot(toPHebdo(larochelle), grep("N", names(larochelle), value=TRUE), ask.leg=FALSE)
  plot(toMois(larochelle), grep("N", names(larochelle), value=TRUE), ask.leg=FALSE)
  # }

[Package Qair version 0.16.4 Index]