plot.rose(Qair)	R Documentation

Affichage de roses de pollution

Description

Permet d'afficher toutes sortes de roses de pollution a partir d'objet de classe 'Qair' et 'rose'

Usage

## S3 method for class 'rose':
plot(x, ..., r, rlim, add = FALSE, thetaPas = 1, density = NULL, col, border, lwd.border = 1, lty.border = 1, at.axis, lwd.axis = 1, lty.axis = 2, col.axis = "grey", label.axis, alpha.axis = 45, cex.axis = 0.7, ray = TRUE, lwd.ray = 0.5 * lwd.border, lty.ray = 1, col.ray = "black", typeRose = c("centre", "zero", "boite"))

rosepol(donnees, formula, breaksVent = 0:36 * 10 - 5, FUN, ...)
## S3 method for class 'Qair':
rosepol(donnees, formula, breaksVent = 0:36 * 10 - 5, FUN, ...)

Arguments

x 	 Un objet de classe 'rose' dont l'utilisateur veut obtenir la representation. Si cet objet n'est pas aggrege, la fonction mean sera utilisee par defaut.

... 	 non utilise. present pour compatibilite.

r 	 Un vecteur de 'character' correspondant aux donnees de rose dont on veut avoir la representation. Par defaut, toutes les donnees de 'rose' seront consecutivement affichees.

rlim 	 Un vecteur de 2 elements, ou une liste de vecteurs de 2 elements. Les 2 elements de chaque vecteur indique la valeur au centre du graphique et l'extension maximale de ce dernier.

add 	 logical, indique si la representation de la rose des pollutions doit etre ajoutee a un graphique existant. (Par defaut un nouveau graphique est cree.)

thetaPas 	 numeric indiquant, en degres, la finesse du trace de la rose de pollution.

density 	 Indique la methode de remplissage des secteurs de la rose de pollution. Une valeur NULL ou de -1 induit le remplissage des secteurs, une valeur de 0 implique que seuls les bordes seront traces, et une valeur positive correspondra a un hachurage. (cf. pour plus de details : polygon.)

col 	 Vecteur de couleurs servant a remplir les secteurs. Si plusieurs roses sont superposees (cas des roses de quantiles), ce vecteur sera recycle si sa taille et le nombre de roses ne correspondent pas.

border 	 Vecteur de couleurs servant pour la bordure des secteurs. Noir par defaut, ce vecteur prend la valeur de 'col' dans le cas ou density = 0

lwd.border 	 Epaisseur des bordures des secteurs. (cf. plot.default)

lty.border 	 Type des bordures. (cf. plot.default)

at.axis 	 Vecteur de 'numeric' definissant les cercles de graduation. Par defaut, 5 cercles reguliers sont traces.

lwd.axis 	 Epaisseur des cercles de graduation. (cf. plot.default)

lty.axis 	 Type des cercles de gradution. En lignes discontinues par defaut. (cf. plot.default)

col.axis 	 Couleur des cercles de graduation. "grey" par defaut.

label.axis 	 Vecteur de 'character'. Legende sur les cercles de graduation. Par defaut, la graduation avec l'unite correspondant au polluant affiche. (L'unite est recuperee a partir de l'objet 'rose')

alpha.axis 	 Vecteur de 'numeric' correspondant aux angles (en degres) auxquels doivent etre affiches les 'label.axis'.

cex.axis 	 numeric. taille de la police (character expansion) des label.axis. (cf. plot.default)

ray 	 logical. Indique si les secteurs doivent etre visuellement separes par des segments. (si density = 0, n'est pas applicable.)

lwd.ray 	 Epaisseur des segments de separation. (cf. plot.default)

lty.ray 	 Type des segments de separation. (cf. plot.default)

col.ray 	 Couleurs des segments de separation. (cf. plot.default)

typeRose 	 'character' specifiant le type de la rose : 'normal' signifie que les secteurs partent du centre du graphique, 'zero' que les secteurs partent de la graduation 0, 'boite' signifie que des 'cases' sont traces. Ce dernier est principalement utilise pour la representation de quantile.

donnees 	 objet de type 'Qair' contenant les donnees a representer, ainsi que des donnees de directions et de vitesses de vent.

formula 	 de la forme ~dv+vv avec dv le nom de la colonne des directions de vent et vv celui de la colonne des vitesses de vent. (cf. calcRose)

breaksVent 	 un vecteur d'angles (en degre) correspondant aux separations entre les secteurs souhaites sur la rose de pollution. Les valeurs entrees sont automatiquement ramenees entre 0 et 360 degre. (cf. calcRose)

FUN 	 la fonction a utiliser pour l'aggregation des donnees. Si 'FUN' n'est pas precisee lorsque calcRose est appliquee a un objet 'rose', la fonction 'mean' est appliquee par defaut avec l'argument na.rm=T. (cf. calcRose)

... 	 Permet de preciser des arugments specifiques a la fonction FUN. (cf. calcRose), ou a la fonction 'plot'. (pour rosepol et rosevent.)

Details

Cette fonction s'appuie sur la classe 'rose', donc des donnees reparties selon les directions de vent, pour representer des roses de pollution. Les roses peuvent etre des roses de moyennes par secteur, de quantiles, de min, de max, de nombre de donnees, etc.

See Also

calcRose, plot

Examples

  data(larochelle)
  rose <- calcRose(larochelle, ~DV_ATM+VV_ATM)
  larochelle$diffO3 <- larochelle$O3_VAU - larochelle$O3_AYT
  attributes(larochelle)$unite[names(larochelle) == "diffO3"] <- "µg/m3"

  # 1{
  par(mfrow=c(2, 2))
  plot(rose["O3_VER"])
  plot(calcRose(larochelle, ~DV_ATM+VV_ATM, FUN="max", na.rm=TRUE)["O3_VER"])
  plot(calcRose(larochelle, ~DV_ATM+VV_ATM, FUN="quantile", na.rm=TRUE), r="DV_ATM", rlim=c(160, 360), typeRose="boite")
  plot(calcRose(larochelle, ~DV_ATM+VV_ATM, FUN="mean", na.rm=TRUE)["diffO3"], col="skyblue", typeRose="zero")
  # }

[Package Qair version 0.16.4 Index]