print.rose(Qair)	R Documentation

Printing 'rose'

Description

Print a 'rose'.

Usage

## S3 method for class 'rose':
print(x, ...)

Arguments

x 	 un objet de classe 'rose' dont on veut afficher les caracteristiques principale.

... 	 Arguments to be passed to or from other methods.

Note

Si l'utilisateur souhaite acceder au contenu reel de la 'rose', il suffit d'utiliser la comande as.list(rose).

See Also

print, as.list.rose

Examples

  data(larochelle)
  print(calcRose(larochelle, ~DV_ATM+VV_ATM))

[Package Qair version 0.16.0 Index]