rosevent(Qair)	R Documentation

Affichage de roses de vent

Description

Permet d'afficher des roses de vent à partir d'objet 'Qair'

Usage

rosevent(donnees, formula, breaksVent, vlim, leg=locator(1), col, ...)
## S3 method for class 'Qair':
rosevent(donnees, formula, breaksVent, vlim, leg=locator(1), col, ...)

calcRoseVent(vv, vlim, tot)

Arguments

donnees 	 objet de type 'Qair' contenant les donnees a representer, ainsi que des donnees de directions et de vitesses de vent.

formula 	 de la forme ~dv+vv avec dv le nom de la colonne des directions de vent et vv celui de la colonne des vitesses de vent. (cf. calcRose)

breaksVent 	 un vecteur d'angles (en degre) correspondant aux separations entre les secteurs souhaites sur la rose de pollution. Les valeurs entrees sont automatiquement ramenees entre 0 et 360 degre. (cf. calcRose)

vlim 	 vecteur contenant des numerics. Definit les classes de vitesses de vents pour la rose de vent. la valeur prise par defaut des classes est calculee a partir des centiles 25, 50, 75 et 100. (donc 4 classes de vitesses de vent.)

leg 	 liste contenant un element 'x' et un 'element 'y' specifiant les coordonnees ou doit apparaitre la legende. Par defaut, la fonction demande a l'utilisateur de 'cliquer cette position'. Si cette valeur est NULL, la legende n'est pas affichee.

col 	 Vecteur de couleurs servant a remplir les secteurs. Si plusieurs roses sont superposees (cas des roses de quantiles), ce vecteur sera recycle si sa taille et le nombre de roses ne correspondent pas.

... 	 Permet de preciser des arugments specifiques a la fonction 'plot.rose'. (plot.rose)

Details

Cette fonction s'appuie sur la classe 'rose', donc des donnees reparties selon les directions de vent, pour representer des roses de vent.

La fonction calcRoseVent est utilisee en interne pour le calcul de la rose des vents.

See Also

calcRose, plot, plot.rose

Examples

  data(larochelle)
  rosevent(larochelle, ~DV_ATM+VV_ATM, leg=NULL)

[Package Qair version 0.16.4 Index]