toAn(Qair)	R Documentation

Convertir des donnees 'Qair' d'une base temporelle a une autre.

Description

Ces differentes fonctions permettent de realiser des conversions entre les differents sous-types de la classe 'Qair'. (toujours de la plue petite unite temporelle vers la plus grande)

Usage

toHeure(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'qh':
toHeure(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'Qair':
toHeure(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)

toJour(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'qh':
toJour(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'heure':
toJour(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'Qair':
toJour(donnees, pc = 0.75, prec = NA, FUN = "mean", nbIndiv, ...)

toHebdo(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'qh':
toHebdo(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'heure':
toHebdo(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'jour':
toHebdo(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'Qair':
toHebdo(donnees, pc = 0.75, prec = NA, FUN = "mean", nbIndiv, ...)

toMois(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'qh':
toMois(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'heure':
toMois(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'jour':
toMois(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'Qair':
toMois(donnees, pc = 0.75, prec = NA, FUN = "mean", nbIndiv, ...)

toAn(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'qh':
toAn(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'heure':
toAn(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'jour':
toAn(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'hebdo':
toAn(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'mois':
toAn(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'Qair':
toAn(donnees, pc = 0.75, prec = NA, FUN = "mean", nbIndiv, ...)

Arguments

donnees 	 Objet de typ 'Qair' a convertir.

pc 	 valeur comprise entre 0 et 1, indiquant le pourcentage de representativite pour qu'une valeur soit conservee. Par exemple, avec pc=0.75, une moyenne horaire ne sera calculee a partir de valeurs qh que si au moins 3 quarts d'heure sont connus pour l'heure dite.

prec 	 numeric ou vecteur de numeric indiquant la precision souhaitee du resultat. Les regles des arrondis applicables dans la qualite de l'air sont utilisees ici. (1 -> un chiffre apres la virgule, 2 -> deux chiffres, etc.)

FUN 	 La fonction a appliquer. Par defaut les moyennes sont calculees, mais il est possible de calculer d'autres indicateurs. (valeur horaire max. sur une journee, etc.)

... 	 Arguments supplementaires necessaires au bon fonctionnement de la fonction FUN. Dans le cas d'un appel a la fonction mean, l'argument na.rm est mis a TRUE.

nbIndiv 	 Parametre precisant de quel sous-type est le 'Qair' a convertir. (L'utilisateur n'est pas encourage a utiliser les fonctions utilisant cet argument.)

Details

Les methodes generales pour la classe 'Qair', ne doivent normalement pas etre utilisees par l'utilisateur, elles servent uniquement dans le cas d'appel aux autres methodes.

Value

Renvoie un objet de classe 'Qair', dont le sous-type est celui precise par l'appel de la fonction.

See Also

toPJour, toPMois, toPAn

Examples

  data(larochelle)

  # toHeure1
  toHeure(larochelle["07-01-01", "07-01-10"])

  # toJour1
  toJour(larochelle["07-01-01", "07-02-01"])
  # toJour2
  toJour(toHeure(larochelle["07-01-01", "07-02-01"]))

  # toHebdo1
  toHebdo(larochelle)
  # toHebdo2
  toHebdo(toHeure(larochelle))
  # toHebdo3
  toHebdo(toJour(larochelle))

  # toMois1
  toMois(larochelle)
  # toMois2
  toMois(toHeure(larochelle))
  # toMois3
  toMois(toJour(larochelle))

  # toAn1
  toAn(larochelle)
  # toAn2
  toAn(toHeure(larochelle))
  # toAn3
  toAn(toJour(larochelle))
  # toAn4
  toAn(toHebdo(larochelle))
  # toAn5
  toAn(toMois(larochelle))

[Package Qair version 0.16.0 Index]