toPAn(Qair)	R Documentation

Calculer des profils moyens (ou autre) journaliers mensuels annuels

Description

Permet de calculer des profils de comportements sur une periode (jour, semaine, mois, annee) a partir de donnees 'Qair'.

Usage

toPJour(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'qh':
toPJour(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'heure':
toPJour(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'Qair':
toPJour(donnees, pc = 0.75, prec = NA, FUN = "mean", heures, ...)

toPHebdo(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'qh':
toPHebdo(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'heure':
toPHebdo(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'jour':
toPHebdo(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'Qair':
toPHebdo(donnees, pc = 0.75, prec = NA, FUN = "mean", jours, ...)

toPMois(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'qh':
toPMois(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'heure':
toPMois(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'jour':
toPMois(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'Qair':
toPMois(donnees, pc = 0.75, prec = NA, FUN = "mean", jours, ...)

toPAn(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'qh':
toPAn(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'heure':
toPAn(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'jour':
toPAn(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'mois':
toPAn(donnees, pc = 0.75, prec = NA, FUN = "mean", ...)
## S3 method for class 'Qair':
toPAn(donnees, pc = 0.75, prec = NA, FUN = "mean", mois, ...)

Arguments

donnees 	 Objet de typ 'Qair' sur lequel on veut calculer un profil.

pc 	 valeur comprise entre 0 et 1, indiquant le pourcentage de representativite pour qu'une valeur soit conservee.

prec 	 numeric ou vecteur de numeric indiquant la precision souhaitee du resultat. Les regles des arrondis applicables dans la qualite de l'air sont utilisees ici. (1 -> un chiffre apres la virgule, 2 -> deux chiffres, etc.)

FUN 	 La fonction a appliquer. Par defaut les moyennes sont calculees, mais il est possible de calculer d'autres indicateurs. (valeur max. chaque sur une semaine, etc.)

... 	 Arguments supplementaires necessaires au bon fonctionnement de la fonction FUN. Dans le cas d'un appel a la fonction mean, l'argument na.rm est mis a TRUE.

heures 	 Parametre correspondant aux heures de la journee. Argument a ne pas utiliser puisque dans une fonction interne.

jours 	 Parametre correspondant aux jours du mois. Argument a ne pas utiliser puisque dans une fonction interne.

mois 	 Parametre correspondant aux mois de l'annee. Argument a ne pas utiliser puisque dans une fonction interne.

Details

Les profils journaliers obtenus seront toujours sur une base horaire, meme si les donnees d'entree sont en quart-horaire. Les profils hebdomadaires seront toujours sur une base journaliere, meme si les donnees d'entree sont en quart-horaire ou horaire. Les profils mensuels obtenus seront toujours sur une base journaliere, meme si les donnees d'entree sont en quart-horaire ou horaire. Les profils annuels obtenus seront toujours sur une base mensuelle, meme si les donnees d'entree sont en quart-horaire, horaire ou journalier.

Value

Un objet de la classe 'Qair', dont la sous-classe sera 'pjour', 'pmois', 'phebdo' ou 'pan' selon les cas.

See Also

toAn

Examples

  data(larochelle)
  # toPJour1
  toPJour(larochelle)
  # toPJour2
  toPJour(toHeure(larochelle))

  # toPHebdo1
  toPHebdo(larochelle)
  # toPHebdo2
  toPHebdo(toHeure(larochelle))
  # toPHebdo3
  toPHebdo(toJour(larochelle))

  # toPMois1
  toPMois(larochelle)
  # toPMois2
  toPMois(toHeure(larochelle))
  # toPMois3
  toPMois(toJour(larochelle))

  # toPAn1
  toPAn(larochelle)
  # toPAn2
  toPAn(toHeure(larochelle))
  # toPAn3
  toPAn(toJour(larochelle))
  # toPAn4
  toPAn(toMois(larochelle))

[Package Qair version 0.16.0 Index]