[.Qair(Qair)	R Documentation

Extract or replace part of 'Qair' extract part of 'rose'

Description

Permet d'acceder a des sous-groupes d'un objet 'Qair' ou 'rose, et de les remplacer dans le cas des objets 'Qair'.

Usage

## S3 method for class 'Qair':
x[i, j, drop]
\method{$<-}{Qair}(x, name, value)

## S3 method for class 'rose':
x[i, j, drop]

Arguments

x 	 objet de classe 'Qair' ou de classe 'rose'

i 	 voir details

j 	 voir details

drop 	 Uniquement pour la compatibilite des methodes, ne pas utiliser.

name 	 voir details

value 	 voir details

Details

Pour Les objets de classe 'Qair', les methodes d'extractions fontcionnent comme pour les data.frame. (cf. [.data.frame) Pour les objets de classe 'rose', la methode '[' permet d'acceder aux elements de objets 'Qair' qu'elle copntient. Appliquer '[' a un rose revient a l'appliquer a chacun des secteurs qui la compose.

Une posssibilite d'extraction suppplementaire a ete ajoutee : elle permet de faire une selection par date. Ainsi il est possible d'indiquer une date de debut en donnant la valeur a 'i' et une date de fin en donnant la valeur a j. Les dates sont indiquees sous le format suivant "y-m-d" s'il la date est en jour, c("y-m-d", "h:m:s") pour prendreen compte les heures.

Value

Un objet dans la meme classe que celui a qui est applique la selection. La colonne 'date' fait systematiquement partie de l'element extrait.

See Also

merge.Qair, [.data.frame

Examples

[Package Qair version 0.16.0 Index]